Simplify the example
Instead of having a controller and consumer node, the example now uses a
single node for both and presents the volume id and the path of the
device where the volume gets attached.

In the readme we remove the inventory section from the example to
simplify it, and recommend running it in a VM to avoid installing
packages on our node.