Update RDO repository to Stein
When we updated the repositories from OSP to 14 we forgot to update RDO
repository to Stein as well.

This patch updates it and also changes the LVM iSCSI configuration in
the docs and the example since we no longer use iscsi_protocol and
iscsi_helper and use target_protocol and target_helper instead.