Speed up second call to extend
Re-running a task that extends a volume is slow on the consumer side
call because we call os-brick's extend method unconditionally.

This patch changes this behavior and will only call os-brick if the
volumes has not already been extended, thus improving the speed of a
second call to extend a volume that has already been extended an the new
size is already known in the system.